Attempt at smarter crawler.
The first problem is that vintages.com and the LCBO On-The-Go API are
only updated daily for inventory, the product catalog data seems to be
updated infrequently, bi-weekly at best. The second problem is that
LCBO.com is slow as molasses, like 10-12 RPM max slow so crawling it in
its entirety every day is not possible.

This commit attempts to be a bit smarter about the products that are
crawled from LCBO.com in the following ways:

1. It only crawls LCBO.com if the product doesn’t exist in OTG.
2. It only crawls LCBO.com if the OTG data is differing.

Hopefully this helps a bit and keeps the data as consistent as possible.